Initial commit
Flexbox based layout for the loop. Added support for Infinite scroll. The advantage is obviously infinite scroll, but the disadvantage is that once you hit the back button after clicking a link in the loop, the infinite scroll does not remember the position you were in. This makes this branch uninteresting for further development.